separated javascript to own file for itemdrop
