--- globals/DROPTYPE.js
@@ -1,9 +1,0 @@
-const DROPTYPE = {
-    0:  'Normal',
-    1:  'Grouped',
-    2:  'Stolen',
-    // 3
-    4:  'Despoiled'
-}
-
-export default DROPTYPE
