+++ globals/droptype.json
@@ -1,0 +1,6 @@
+{
+    "0": "Normal",
+    "1":  "Grouped",
+    "2":  "Stolen",
+    "4": "Despoiled"
+}
